pyproject.toml
[project]
name = "invariant-gfx"
version = "0.2.0.dev0"
description = "A deterministic, functional graphics pipeline built on Invariant. Build icons, badges, dynamic UI components, and data visualizations by composing reusable pipeline parts in a DAG-based system."
authors = [
    {name = "Kaj Siebert", email = "kaj@k-si.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10,<4.0"
keywords = [
    "graphics",
    "dag",
    "invariant",
    "rendering",
    "pil",
    "pillow",
    "svg",
    "composition",
    "deterministic",
    "cacheable",
    "layout",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    "invariant-core>=0.2.0",
    "justmytype>=0.3.0,<1.0.0",
    "justmyresource>=1.1.1,<2.0.0",
    "Pillow>=12.1.1,<13.0.0",
    "cairosvg>=2.8.2,<3.0.0",
]

[project.urls]
Homepage = "https://github.com/kws/invariant-gfx"
Repository = "https://github.com/kws/invariant-gfx"
Documentation = "https://github.com/kws/invariant-gfx"
"Bug Tracker" = "https://github.com/kws/invariant-gfx/issues"

[tool.uv.sources]
invariant-core = {path = "../invariant", editable = true}

[tool.hatch.build.targets.wheel]
packages = ["src/invariant_gfx"]

[dependency-groups]
dev = [
    "ruff>=0.15.1",
    "pytest>=9.0.2",
    "pre-commit>=3.8.0",
    "justmytype>=0.3.0",
    "justmyresource[icons]>=1.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
